Implemented dynamic pokemon palettes.
The variant attribute consists of 4 3-bit palette indexes and a sprite
index (currently unused). The variants data defines what parts of the
sprite correspond to which colors of a palette. These can be
mixed-and-matched for a much greater variety of pokemon variants.

(NOTE) The dynamic palette makes use of a couple intermediary variables.
At this point, I don't know if I've implemented them correctly, with
respect to memory allocation. It's unlikely, but there may be memory
leaks somewhere down the line.