Fixed the pokemon palette when viewed in the PC.
This was caused by the dynamic palette in pokemon.c. The palette would
get generated dyanamically, but then the pointer to it is stored in
gStorage->displayMonPalette. It wouldn't get loaded until later, and
during that time, the dynamic palette data would get corrupted with
something else.
The solution was to bypass storing the palette, and simply generate it
immediately before it's loaded, based on specise and variant.