Tried to import platinum sprites.
Ultimately, this failed.
GBA does not support sprites larger than 64x64 (platinum is 80^2), so I tried a bunch of algorithms to downscale/crop without success. Ultimately, it would take too much effort to modify the sprites individually. The final option would be a system of using multiple sprites in-game, which would be obscenely complicated.
I'm preserving this work, but this feature is probably abandoned.